Calculate FOS of footprints

Calculate the FOS(footprints occupancy score) to identify enriched transcriptions factor, and use these transcription factors to identify significant regulatory relationships

Footprints_FOS(Wig_list, Candid, FOS_threshold = 1, trans_wig = FALSE)

Arguments

Wig_list
list object, where each element is the wig_track result of each sample.

Candid
If you follow our our pipline, this parameter should be the second element of the list that generated by get_peaks_genes()

FOS_threshold
numeric, indicating the threshold of footprint occupancy score to identify significant footprints.

trans_wig
logcial, if you use dnase_wig_tracks_both2.py to get wig_list, please set trans_wig parameter as TRUE.

Value

return related transcription factors of footprints with FOS above FOS_threshold

Examples

load(system.file("extdata", "Candid.rda", package = "IReNA"))
load(system.file("extdata", "wig_list.rda", package = "IReNA"))
regulatory_relationships <- Footprints_FOS(wig_list, Candid)

#> [1] 886
#> [1] 329