add pseudtoime and identify DEGs

Add pseudotime to seurat object, and filter differential expressed genes according to pseudotime

add_pseudotime_DEG_filter(
  seurat_object,
  monocle_object,
  DEG = TRUE,
  qvalue = 0.05,
  nce = 0.1,
  ed = 0.1,
  normlize1 = TRUE
)

Arguments

seurat_object
seurat object, which has same cells and genes as monocle_object

monocle_object
monocle object, which include pesudotime and has same cells and genes as seurat_object

DEG
logic, indicating whether filter differential expressed genes for seurat object

qvalue
numeric, indicating q-value to indentify differentially expressed genes

nce
numeric, indicating num cells expressed to indentify differentially expressed genes

ed
numeric, indicating expression difference to indentify differentially expressed genes

normlize1
logic, indicating whether normalize the data in seurat object

Value

return a seurat object

Examples

load(system.file("extdata", "test_seurat.rda", package = "IReNA"))
monocle_object = get_pseudotime(test_seurat)

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

add_pseudotime_DEG_filter(seurat_object = test_seurat,monocle_object = monocle_object, DEG = FALSE, normlize1 = FALSE)

#> Loading required package: SeuratObject

#> An object of class Seurat 
#> 100 features across 10 samples within 1 assay 
#> Active assay: RNA (100 features, 0 variable features)

#add_pseudotime_DEG_filter(seurat_object = test_seurat,monocle_object = monocle_object, DEG = TRUE, qvalue = 0.001, nce = 0.1, ed = 0.1)