get somoothbybin

Divide cells into bins across pseudotime and return expression profile

get_SmoothByBin_PseudotimeExp(
  seurat_object,
  FC = TRUE,
  Bin = 50,
  FcType = "Q95"
)

Arguments

seurat_object
seurat object, where meta.data should contain Pseudotime

FC
logic, indicating whether to add FoldChangeQ95 to the first column, default is TRUE

Bin
numeric, indicating the numbers of bin which divide the pseudotime, default is 50

FcType
'Q95' or 'Q90', FoldChange threshold, default is 'Q95'

Value

return a expression profile

Examples

load(system.file("extdata", "test_seurat.rda", package = "IReNA"))
monocle_object = get_pseudotime(test_seurat)

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

seurat_object_pseudotime=add_pseudotime_DEG_filter(seurat_object = test_seurat,monocle_object = monocle_object, DEG = FALSE, normlize1 = FALSE)
get_SmoothByBin_PseudotimeExp(seurat_object_pseudotime, Bin = 50, FcType = "Q95")

#>          FoldChangeQ95    SmExp1 SmExp2 SmExp3 SmExp4    SmExp5    SmExp6
#> TCEB3        1.1626456  1.811529      0      0      0  2.903650 -0.241112
#> CLK1         2.2202732  2.952432      0      0      0  2.067283  0.851007
#> MATR3        1.0517153  0.216634      0      0      0  0.169355  0.675605
#> AKAP11       1.0030214  1.413757      0      0      0 -0.209577  0.188259
#> HSF2         0.9115132 -0.766082      0      0      0  0.575873  0.382621
#> TBPL1        2.1765200  1.978140      0      0      0 -0.849110  0.206193
#> MAGEC2       2.1883540  3.094644      0      0      0  1.057962  1.670331
#> DDX20        1.9273114  0.899471      0      0      0  2.900351  0.052574
#> ZBTB11       1.4396683  0.836996      0      0      0 -1.096010  1.709095
#> ARFGEF1      1.2557452 -0.731054      0      0      0 -0.833417  0.457326
#> FGFR3        0.8748540  1.728555      0      0      0  2.666341  0.660915
#> IKBKAP       0.8674748 -0.295984      0      0      0  1.335325  0.612391
#> ASNS         1.0237484 -0.093569      0      0      0 -1.136460 -0.015380
#> TCF3         0.7206952 -0.113853      0      0      0  0.417285  0.301670
#> PDCD2        0.8184018  2.136827      0      0      0 -0.795486 -0.102003
#> SMARCE1      1.1606980  1.858918      0      0      0  1.526533  0.626717
#> LMAN1        0.7437254 -0.463706      0      0      0 -0.632412  0.341521
#> SLC25A3      2.3185367  1.838262      0      0      0 -2.108613  0.972705
#> REXO2        1.8688650  1.961321      0      0      0 -1.011443  1.609912
#> DDX43        1.3917991  1.091933      0      0      0 -0.987679  2.974232
#> BZW1         0.9274516  1.159104      0      0      0  0.687968  0.702777
#> CCNT2        1.3614657 -0.925197      0      0      0  2.215443  1.157394
#> REST         1.6258552 -1.187611      0      0      0  1.760707  0.430161
#> MYNN         0.9110961 -0.940405      0      0      0 -0.822210  0.541924
#> ZNF213       1.4079548  2.722484      0      0      0 -0.387222  0.889048
#> AKR1B1       1.2007855  0.860378      0      0      0 -0.969913 -0.175949
#> SF3B2        1.3699955  1.435055      0      0      0 -1.397534  1.150430
#> PABPC4       0.6961724 -1.149397      0      0      0  1.737002  0.295045
#> CDV3         1.8927238  1.698525      0      0      0  0.740644  2.035325
#> HECTD1       1.1300902 -0.363614      0      0      0  0.409322  0.412867
#> TDRD1        0.7413039 -0.845372      0      0      0 -0.098279 -0.368373
#> BRPF3        1.0176460  1.082157      0      0      0  0.297764 -0.523194
#> TMEM14A      1.4585245  1.677775      0      0      0  1.824793  0.216201
#> MAGEB2       1.5923168 -0.873037      0      0      0  0.229205  1.406446
#> SLC25A1      1.2758001 -0.582323      0      0      0  1.340715 -0.938590
#> TFIP11       1.3372495 -0.831131      0      0      0  3.816271  0.087126
#> FBXO7        0.8133482 -1.355831      0      0      0 -1.840932  0.698085
#> RPL3         0.6350649  0.007085      0      0      0  0.492332  0.966136
#> EIF5         1.0633053  0.369695      0      0      0  1.291224  0.379814
#> APEX1        1.9252233  1.995876      0      0      0 -1.329457  1.838870
#> RABGGTA      1.3217720  2.530625      0      0      0 -1.052686 -1.199675
#> SALL4        1.8510754 -0.531039      0      0      0 -0.513216  0.168027
#> ADNP         1.3997103  0.539037      0      0      0  1.412761  0.025592
#> PSMA7        1.9840513 -1.457991      0      0      0 -1.404053  1.515062
#> MRGBP        2.0284468 -0.847836      0      0      0 -0.924341 -0.104035
#> ASB9         1.1804455 -0.652688      0      0      0  0.876860 -0.832817
#> RBBP7        1.1335289 -0.852040      0      0      0  1.563133  0.032325
#> USP11        1.3841110 -1.070883      0      0      0  1.711682 -0.679264
#> CHRAC1       1.4999618  1.990000      0      0      0  0.778377 -0.701736
#> LIN7B        3.3093654  2.339291      0      0      0  3.228726 -0.852626
#> GPI          0.9004190  0.442538      0      0      0  2.931245  0.098404
#> GARS         2.1192648 -1.379382      0      0      0 -0.028951 -0.254503
#> SETX         1.1518269  1.156214      0      0      0 -1.773670  0.327675
#> ZFAND5       1.6722888 -1.772988      0      0      0  0.742914  0.786241
#> TFAM         1.4745440  1.378383      0      0      0 -1.201614  1.019135
#> RPL19        1.1752475  1.144162      0      0      0 -1.002385  0.475400
#> FTSJ3        1.7398641 -1.025487      0      0      0  1.481455  1.398674
#> VAT1         0.6414251  0.718188      0      0      0 -0.707101  0.073308
#> UNC119       1.1618519  3.203493      0      0      0 -0.651965 -0.849244
#> GRPEL1       1.0620029 -1.141728      0      0      0  0.526299  0.905991
#> HSPA8        0.8179233  1.289966      0      0      0  1.376575  0.484242
#> DDX6         1.0582901  0.014872      0      0      0  1.133536 -0.280582
#> CBL          0.9388516  0.869210      0      0      0 -0.615065  0.524956
#> CAPRIN2      0.8699877  1.242262      0      0      0 -1.628751  0.350807
#> KRR1         1.0654641 -0.571192      0      0      0 -1.049414 -0.793773
#> TPI1         1.3680889 -2.244820      0      0      0 -2.236556  1.367969
#> TDP2         1.4081651 -0.549395      0      0      0  3.220332 -0.989628
#> ZNF451       0.8705452  0.385924      0      0      0 -0.905125  0.936572
#> RPS12        0.9234476  0.609586      0      0      0  1.178262  0.565552
#> ARRDC3       2.7290625  1.200204      0      0      0 -0.229131  1.862336
#> RARS         2.2185792  0.316525      0      0      0 -1.444145  0.990130
#> ACTR8        0.8935444  0.918512      0      0      0  2.859873  0.016849
#> EEF1B2       1.0883184 -1.479029      0      0      0  2.210125  0.669750
#> ID2          2.1359978  1.490636      0      0      0  1.874602  3.446614
#> ODC1         1.2509718 -0.469247      0      0      0  3.430825  0.267837
#> CACYBP       1.0449959  0.591551      0      0      0  1.947035  0.713511
#> LRRC42       1.7873228  0.881934      0      0      0 -1.257636 -0.993808
#> PHF13        1.7584311  1.994591      0      0      0  1.118540 -1.128029
#> ADPRHL2      0.9982403 -0.972920      0      0      0  0.220038  0.558429
#> EXOC8        1.9588735  2.115995      0      0      0 -0.814285  1.916838
#> TXNIP        2.5296635  2.056251      0      0      0  2.322213 -1.298322
#> CCNI         0.9281586  0.427908      0      0      0  0.883442  0.268044
#> ALG2         0.7039794 -0.720691      0      0      0 -0.602496  2.731451
#> NEK9         0.6410641 -0.734205      0      0      0 -0.587647  1.148072
#> IRF2BPL      1.1651760  1.186524      0      0      0  0.852462  0.587138
#> TMEM214      2.0113405  2.262996      0      0      0  2.161591 -0.796268
#> YPEL5        2.2211779 -1.908173      0      0      0 -2.211777  1.369870
#> FAM98A       1.7184772 -1.275177      0      0      0  1.271598  0.396545
#> AFTPH        0.8439606  1.311929      0      0      0 -0.565575 -0.002553
#> LGALSL       1.6455596 -0.200775      0      0      0  4.207653  0.252284
#> MRPS14       1.2422318 -0.691051      0      0      0  2.540253  0.917001
#> NFYB         0.9403585 -0.749305      0      0      0 -0.670553 -0.835851
#> DPPA4        1.3491920  1.388218      0      0      0  1.432264  0.389239
#> NAA50        1.1499166 -1.095960      0      0      0  1.144194 -0.221146
#> FYTTD1       1.6710586 -0.946387      0      0      0  1.947828  1.270481
#> NUDT10       0.8184196  0.062854      0      0      0  4.481645  0.434045
#> EXOSC9       2.3697909 -1.759019      0      0      0 -0.334509 -1.340406
#> RAB22A       1.4007751  0.543694      0      0      0  1.639353  0.842499
#> SOX4         1.0497592  0.437487      0      0      0  0.931001  1.146925
#> C17orf53     0.7085820  0.721088      0      0      0 -0.331035  0.312206
#>          SmExp7    SmExp8 SmExp9 SmExp10 SmExp11 SmExp12 SmExp13 SmExp14
#> TCEB3         0 -0.754542      0       0       0       0       0       0
#> CLK1          0  1.225657      0       0       0       0       0       0
#> MATR3         0  1.548587      0       0       0       0       0       0
#> AKAP11        0  1.294035      0       0       0       0       0       0
#> HSF2          0  0.817219      0       0       0       0       0       0
#> TBPL1         0 -0.888747      0       0       0       0       0       0
#> MAGEC2        0  2.971799      0       0       0       0       0       0
#> DDX20         0 -1.215724      0       0       0       0       0       0
#> ZBTB11        0  2.063035      0       0       0       0       0       0
#> ARFGEF1       0 -0.587506      0       0       0       0       0       0
#> FGFR3         0 -0.618336      0       0       0       0       0       0
#> IKBKAP        0 -0.539959      0       0       0       0       0       0
#> ASNS          0 -0.972909      0       0       0       0       0       0
#> TCF3          0  0.621368      0       0       0       0       0       0
#> PDCD2         0 -0.768527      0       0       0       0       0       0
#> SMARCE1       0 -0.191630      0       0       0       0       0       0
#> LMAN1         0  1.903908      0       0       0       0       0       0
#> SLC25A3       0 -1.565844      0       0       0       0       0       0
#> REXO2         0 -0.231419      0       0       0       0       0       0
#> DDX43         0 -0.305432      0       0       0       0       0       0
#> BZW1          0  0.053576      0       0       0       0       0       0
#> CCNT2         0  1.235085      0       0       0       0       0       0
#> REST          0 -0.800580      0       0       0       0       0       0
#> MYNN          0 -0.844988      0       0       0       0       0       0
#> ZNF213        0  2.881940      0       0       0       0       0       0
#> AKR1B1        0  0.207444      0       0       0       0       0       0
#> SF3B2         0  0.903150      0       0       0       0       0       0
#> PABPC4        0  0.376579      0       0       0       0       0       0
#> CDV3          0  1.884267      0       0       0       0       0       0
#> HECTD1        0  0.220248      0       0       0       0       0       0
#> TDRD1         0  0.244928      0       0       0       0       0       0
#> BRPF3         0 -0.618550      0       0       0       0       0       0
#> TMEM14A       0 -0.731613      0       0       0       0       0       0
#> MAGEB2        0  1.744393      0       0       0       0       0       0
#> SLC25A1       0  1.128427      0       0       0       0       0       0
#> TFIP11        0 -0.754419      0       0       0       0       0       0
#> FBXO7         0  1.087943      0       0       0       0       0       0
#> RPL3          0  1.245242      0       0       0       0       0       0
#> EIF5          0  1.532895      0       0       0       0       0       0
#> APEX1         0  1.135242      0       0       0       0       0       0
#> RABGGTA       0 -1.191477      0       0       0       0       0       0
#> SALL4         0 -0.544989      0       0       0       0       0       0
#> ADNP          0 -1.285253      0       0       0       0       0       0
#> PSMA7         0  2.205614      0       0       0       0       0       0
#> MRGBP         0 -0.947119      0       0       0       0       0       0
#> ASB9          0 -0.720332      0       0       0       0       0       0
#> RBBP7         0  1.092731      0       0       0       0       0       0
#> USP11         0  1.238973      0       0       0       0       0       0
#> CHRAC1        0 -0.920467      0       0       0       0       0       0
#> LIN7B         0 -0.910718      0       0       0       0       0       0
#> GPI           0 -0.004585      0       0       0       0       0       0
#> GARS          0 -1.325552      0       0       0       0       0       0
#> SETX          0 -0.318513      0       0       0       0       0       0
#> ZFAND5        0 -0.528574      0       0       0       0       0       0
#> TFAM          0 -1.224392      0       0       0       0       0       0
#> RPL19         0  1.518029      0       0       0       0       0       0
#> FTSJ3         0 -0.930070      0       0       0       0       0       0
#> VAT1          0  1.396068      0       0       0       0       0       0
#> UNC119        0 -0.674743      0       0       0       0       0       0
#> GRPEL1        0  1.189649      0       0       0       0       0       0
#> HSPA8         0 -0.136668      0       0       0       0       0       0
#> DDX6          0  2.060371      0       0       0       0       0       0
#> CBL           0  1.102013      0       0       0       0       0       0
#> CAPRIN2       0 -0.014806      0       0       0       0       0       0
#> KRR1          0 -1.072192      0       0       0       0       0       0
#> TPI1          0  1.866815      0       0       0       0       0       0
#> TDP2          0 -0.919315      0       0       0       0       0       0
#> ZNF451        0 -0.306787      0       0       0       0       0       0
#> RPS12         0  0.958623      0       0       0       0       0       0
#> ARRDC3        0 -0.816905      0       0       0       0       0       0
#> RARS          0 -1.472048      0       0       0       0       0       0
#> ACTR8         0 -0.246964      0       0       0       0       0       0
#> EEF1B2        0  1.180815      0       0       0       0       0       0
#> ID2           0  2.379829      0       0       0       0       0       0
#> ODC1          0  1.501064      0       0       0       0       0       0
#> CACYBP        0  0.441434      0       0       0       0       0       0
#> LRRC42        0 -1.258952      0       0       0       0       0       0
#> PHF13         0  1.273981      0       0       0       0       0       0
#> ADPRHL2       0 -0.877503      0       0       0       0       0       0
#> EXOC8         0 -0.589526      0       0       0       0       0       0
#> TXNIP         0  3.955689      0       0       0       0       0       0
#> CCNI          0  1.288073      0       0       0       0       0       0
#> ALG2          0 -0.625274      0       0       0       0       0       0
#> NEK9          0  1.423857      0       0       0       0       0       0
#> IRF2BPL       0  0.254869      0       0       0       0       0       0
#> TMEM214       0  1.827701      0       0       0       0       0       0
#> YPEL5         0  2.019416      0       0       0       0       0       0
#> FAM98A        0  2.698943      0       0       0       0       0       0
#> AFTPH         0 -0.501514      0       0       0       0       0       0
#> LGALSL        0  4.097433      0       0       0       0       0       0
#> MRPS14        0 -0.727426      0       0       0       0       0       0
#> NFYB          0  0.729966      0       0       0       0       0       0
#> DPPA4         0 -0.557059      0       0       0       0       0       0
#> NAA50         0 -0.510385      0       0       0       0       0       0
#> FYTTD1        0  1.028852      0       0       0       0       0       0
#> NUDT10        0 -0.496424      0       0       0       0       0       0
#> EXOSC9        0  0.663317      0       0       0       0       0       0
#> RAB22A        0 -0.114923      0       0       0       0       0       0
#> SOX4          0  0.552071      0       0       0       0       0       0
#> C17orf53      0  0.629511      0       0       0       0       0       0
#>          SmExp15 SmExp16 SmExp17 SmExp18 SmExp19 SmExp20 SmExp21 SmExp22
#> TCEB3          0       0       0       0       0       0       0       0
#> CLK1           0       0       0       0       0       0       0       0
#> MATR3          0       0       0       0       0       0       0       0
#> AKAP11         0       0       0       0       0       0       0       0
#> HSF2           0       0       0       0       0       0       0       0
#> TBPL1          0       0       0       0       0       0       0       0
#> MAGEC2         0       0       0       0       0       0       0       0
#> DDX20          0       0       0       0       0       0       0       0
#> ZBTB11         0       0       0       0       0       0       0       0
#> ARFGEF1        0       0       0       0       0       0       0       0
#> FGFR3          0       0       0       0       0       0       0       0
#> IKBKAP         0       0       0       0       0       0       0       0
#> ASNS           0       0       0       0       0       0       0       0
#> TCF3           0       0       0       0       0       0       0       0
#> PDCD2          0       0       0       0       0       0       0       0
#> SMARCE1        0       0       0       0       0       0       0       0
#> LMAN1          0       0       0       0       0       0       0       0
#> SLC25A3        0       0       0       0       0       0       0       0
#> REXO2          0       0       0       0       0       0       0       0
#> DDX43          0       0       0       0       0       0       0       0
#> BZW1           0       0       0       0       0       0       0       0
#> CCNT2          0       0       0       0       0       0       0       0
#> REST           0       0       0       0       0       0       0       0
#> MYNN           0       0       0       0       0       0       0       0
#> ZNF213         0       0       0       0       0       0       0       0
#> AKR1B1         0       0       0       0       0       0       0       0
#> SF3B2          0       0       0       0       0       0       0       0
#> PABPC4         0       0       0       0       0       0       0       0
#> CDV3           0       0       0       0       0       0       0       0
#> HECTD1         0       0       0       0       0       0       0       0
#> TDRD1          0       0       0       0       0       0       0       0
#> BRPF3          0       0       0       0       0       0       0       0
#> TMEM14A        0       0       0       0       0       0       0       0
#> MAGEB2         0       0       0       0       0       0       0       0
#> SLC25A1        0       0       0       0       0       0       0       0
#> TFIP11         0       0       0       0       0       0       0       0
#> FBXO7          0       0       0       0       0       0       0       0
#> RPL3           0       0       0       0       0       0       0       0
#> EIF5           0       0       0       0       0       0       0       0
#> APEX1          0       0       0       0       0       0       0       0
#> RABGGTA        0       0       0       0       0       0       0       0
#> SALL4          0       0       0       0       0       0       0       0
#> ADNP           0       0       0       0       0       0       0       0
#> PSMA7          0       0       0       0       0       0       0       0
#> MRGBP          0       0       0       0       0       0       0       0
#> ASB9           0       0       0       0       0       0       0       0
#> RBBP7          0       0       0       0       0       0       0       0
#> USP11          0       0       0       0       0       0       0       0
#> CHRAC1         0       0       0       0       0       0       0       0
#> LIN7B          0       0       0       0       0       0       0       0
#> GPI            0       0       0       0       0       0       0       0
#> GARS           0       0       0       0       0       0       0       0
#> SETX           0       0       0       0       0       0       0       0
#> ZFAND5         0       0       0       0       0       0       0       0
#> TFAM           0       0       0       0       0       0       0       0
#> RPL19          0       0       0       0       0       0       0       0
#> FTSJ3          0       0       0       0       0       0       0       0
#> VAT1           0       0       0       0       0       0       0       0
#> UNC119         0       0       0       0       0       0       0       0
#> GRPEL1         0       0       0       0       0       0       0       0
#> HSPA8          0       0       0       0       0       0       0       0
#> DDX6           0       0       0       0       0       0       0       0
#> CBL            0       0       0       0       0       0       0       0
#> CAPRIN2        0       0       0       0       0       0       0       0
#> KRR1           0       0       0       0       0       0       0       0
#> TPI1           0       0       0       0       0       0       0       0
#> TDP2           0       0       0       0       0       0       0       0
#> ZNF451         0       0       0       0       0       0       0       0
#> RPS12          0       0       0       0       0       0       0       0
#> ARRDC3         0       0       0       0       0       0       0       0
#> RARS           0       0       0       0       0       0       0       0
#> ACTR8          0       0       0       0       0       0       0       0
#> EEF1B2         0       0       0       0       0       0       0       0
#> ID2            0       0       0       0       0       0       0       0
#> ODC1           0       0       0       0       0       0       0       0
#> CACYBP         0       0       0       0       0       0       0       0
#> LRRC42         0       0       0       0       0       0       0       0
#> PHF13          0       0       0       0       0       0       0       0
#> ADPRHL2        0       0       0       0       0       0       0       0
#> EXOC8          0       0       0       0       0       0       0       0
#> TXNIP          0       0       0       0       0       0       0       0
#> CCNI           0       0       0       0       0       0       0       0
#> ALG2           0       0       0       0       0       0       0       0
#> NEK9           0       0       0       0       0       0       0       0
#> IRF2BPL        0       0       0       0       0       0       0       0
#> TMEM214        0       0       0       0       0       0       0       0
#> YPEL5          0       0       0       0       0       0       0       0
#> FAM98A         0       0       0       0       0       0       0       0
#> AFTPH          0       0       0       0       0       0       0       0
#> LGALSL         0       0       0       0       0       0       0       0
#> MRPS14         0       0       0       0       0       0       0       0
#> NFYB           0       0       0       0       0       0       0       0
#> DPPA4          0       0       0       0       0       0       0       0
#> NAA50          0       0       0       0       0       0       0       0
#> FYTTD1         0       0       0       0       0       0       0       0
#> NUDT10         0       0       0       0       0       0       0       0
#> EXOSC9         0       0       0       0       0       0       0       0
#> RAB22A         0       0       0       0       0       0       0       0
#> SOX4           0       0       0       0       0       0       0       0
#> C17orf53       0       0       0       0       0       0       0       0
#>          SmExp23 SmExp24 SmExp25   SmExp26 SmExp27 SmExp28 SmExp29 SmExp30
#> TCEB3          0       0       0  0.428017       0       0       0       0
#> CLK1           0       0       0  1.119202       0       0       0       0
#> MATR3          0       0       0  0.304175       0       0       0       0
#> AKAP11         0       0       0  0.393041       0       0       0       0
#> HSF2           0       0       0  0.988663       0       0       0       0
#> TBPL1          0       0       0  0.911181       0       0       0       0
#> MAGEC2         0       0       0 -0.438945       0       0       0       0
#> DDX20          0       0       0  0.008657       0       0       0       0
#> ZBTB11         0       0       0 -0.589084       0       0       0       0
#> ARFGEF1        0       0       0  0.856890       0       0       0       0
#> FGFR3          0       0       0 -0.633933       0       0       0       0
#> IKBKAP         0       0       0 -0.848410       0       0       0       0
#> ASNS           0       0       0 -1.225917       0       0       0       0
#> TCF3           0       0       0  1.234513       0       0       0       0
#> PDCD2          0       0       0 -0.764011       0       0       0       0
#> SMARCE1        0       0       0  0.168903       0       0       0       0
#> LMAN1          0       0       0 -0.834800       0       0       0       0
#> SLC25A3        0       0       0  1.531710       0       0       0       0
#> REXO2          0       0       0 -0.949155       0       0       0       0
#> DDX43          0       0       0  0.245707       0       0       0       0
#> BZW1           0       0       0  1.097197       0       0       0       0
#> CCNT2          0       0       0  0.641991       0       0       0       0
#> REST           0       0       0 -1.544041       0       0       0       0
#> MYNN           0       0       0 -0.870661       0       0       0       0
#> ZNF213         0       0       0  1.001036       0       0       0       0
#> AKR1B1         0       0       0 -0.466972       0       0       0       0
#> SF3B2          0       0       0  0.503522       0       0       0       0
#> PABPC4         0       0       0 -1.714782       0       0       0       0
#> CDV3           0       0       0 -1.249694       0       0       0       0
#> HECTD1         0       0       0  1.119132       0       0       0       0
#> TDRD1          0       0       0  2.411941       0       0       0       0
#> BRPF3          0       0       0  1.216082       0       0       0       0
#> TMEM14A        0       0       0  1.512982       0       0       0       0
#> MAGEB2         0       0       0  0.758321       0       0       0       0
#> SLC25A1        0       0       0 -0.360801       0       0       0       0
#> TFIP11         0       0       0 -0.168776       0       0       0       0
#> FBXO7          0       0       0  0.461111       0       0       0       0
#> RPL3           0       0       0  0.625034       0       0       0       0
#> EIF5           0       0       0  2.367576       0       0       0       0
#> APEX1          0       0       0  2.067867       0       0       0       0
#> RABGGTA        0       0       0  3.529049       0       0       0       0
#> SALL4          0       0       0 -0.318848       0       0       0       0
#> ADNP           0       0       0  0.826368       0       0       0       0
#> PSMA7          0       0       0 -0.255352       0       0       0       0
#> MRGBP          0       0       0  2.070968       0       0       0       0
#> ASB9           0       0       0  0.829485       0       0       0       0
#> RBBP7          0       0       0 -1.564745       0       0       0       0
#> USP11          0       0       0  0.671848       0       0       0       0
#> CHRAC1         0       0       0 -0.887228       0       0       0       0
#> LIN7B          0       0       0 -1.078611       0       0       0       0
#> GPI            0       0       0  1.275049       0       0       0       0
#> GARS           0       0       0  1.892098       0       0       0       0
#> SETX           0       0       0 -0.916838       0       0       0       0
#> ZFAND5         0       0       0  1.950349       0       0       0       0
#> TFAM           0       0       0  1.923670       0       0       0       0
#> RPL19          0       0       0  0.700449       0       0       0       0
#> FTSJ3          0       0       0  3.163156       0       0       0       0
#> VAT1           0       0       0 -0.471556       0       0       0       0
#> UNC119         0       0       0 -0.733023       0       0       0       0
#> GRPEL1         0       0       0  2.585085       0       0       0       0
#> HSPA8          0       0       0  0.785042       0       0       0       0
#> DDX6           0       0       0  1.392696       0       0       0       0
#> CBL            0       0       0  0.134946       0       0       0       0
#> CAPRIN2        0       0       0 -0.419078       0       0       0       0
#> KRR1           0       0       0 -1.057241       0       0       0       0
#> TPI1           0       0       0  0.858072       0       0       0       0
#> TDP2           0       0       0 -1.133387       0       0       0       0
#> ZNF451         0       0       0  1.032467       0       0       0       0
#> RPS12          0       0       0  0.855420       0       0       0       0
#> ARRDC3         0       0       0  2.701375       0       0       0       0
#> RARS           0       0       0  1.625312       0       0       0       0
#> ACTR8          0       0       0  0.893261       0       0       0       0
#> EEF1B2         0       0       0 -2.585081       0       0       0       0
#> ID2            0       0       0 -0.730785       0       0       0       0
#> ODC1           0       0       0 -0.041736       0       0       0       0
#> CACYBP         0       0       0 -0.687358       0       0       0       0
#> LRRC42         0       0       0  1.419430       0       0       0       0
#> PHF13          0       0       0  1.266565       0       0       0       0
#> ADPRHL2        0       0       0 -0.830833       0       0       0       0
#> EXOC8          0       0       0 -0.266895       0       0       0       0
#> TXNIP          0       0       0  2.319025       0       0       0       0
#> CCNI           0       0       0  0.964745       0       0       0       0
#> ALG2           0       0       0 -0.683554       0       0       0       0
#> NEK9           0       0       0 -0.684769       0       0       0       0
#> IRF2BPL        0       0       0  2.026757       0       0       0       0
#> TMEM214        0       0       0  2.298799       0       0       0       0
#> YPEL5          0       0       0  0.929454       0       0       0       0
#> FAM98A         0       0       0  0.172113       0       0       0       0
#> AFTPH          0       0       0  0.693587       0       0       0       0
#> LGALSL         0       0       0 -0.569829       0       0       0       0
#> MRPS14         0       0       0  0.059609       0       0       0       0
#> NFYB           0       0       0 -0.409605       0       0       0       0
#> DPPA4          0       0       0 -0.615339       0       0       0       0
#> NAA50          0       0       0 -0.404182       0       0       0       0
#> FYTTD1         0       0       0  1.396603       0       0       0       0
#> NUDT10         0       0       0  1.680731       0       0       0       0
#> EXOSC9         0       0       0  2.205048       0       0       0       0
#> RAB22A         0       0       0  1.059581       0       0       0       0
#> SOX4           0       0       0  0.028914       0       0       0       0
#> C17orf53       0       0       0 -0.046522       0       0       0       0
#>          SmExp31 SmExp32 SmExp33 SmExp34 SmExp35 SmExp36 SmExp37 SmExp38
#> TCEB3          0       0       0       0       0       0       0       0
#> CLK1           0       0       0       0       0       0       0       0
#> MATR3          0       0       0       0       0       0       0       0
#> AKAP11         0       0       0       0       0       0       0       0
#> HSF2           0       0       0       0       0       0       0       0
#> TBPL1          0       0       0       0       0       0       0       0
#> MAGEC2         0       0       0       0       0       0       0       0
#> DDX20          0       0       0       0       0       0       0       0
#> ZBTB11         0       0       0       0       0       0       0       0
#> ARFGEF1        0       0       0       0       0       0       0       0
#> FGFR3          0       0       0       0       0       0       0       0
#> IKBKAP         0       0       0       0       0       0       0       0
#> ASNS           0       0       0       0       0       0       0       0
#> TCF3           0       0       0       0       0       0       0       0
#> PDCD2          0       0       0       0       0       0       0       0
#> SMARCE1        0       0       0       0       0       0       0       0
#> LMAN1          0       0       0       0       0       0       0       0
#> SLC25A3        0       0       0       0       0       0       0       0
#> REXO2          0       0       0       0       0       0       0       0
#> DDX43          0       0       0       0       0       0       0       0
#> BZW1           0       0       0       0       0       0       0       0
#> CCNT2          0       0       0       0       0       0       0       0
#> REST           0       0       0       0       0       0       0       0
#> MYNN           0       0       0       0       0       0       0       0
#> ZNF213         0       0       0       0       0       0       0       0
#> AKR1B1         0       0       0       0       0       0       0       0
#> SF3B2          0       0       0       0       0       0       0       0
#> PABPC4         0       0       0       0       0       0       0       0
#> CDV3           0       0       0       0       0       0       0       0
#> HECTD1         0       0       0       0       0       0       0       0
#> TDRD1          0       0       0       0       0       0       0       0
#> BRPF3          0       0       0       0       0       0       0       0
#> TMEM14A        0       0       0       0       0       0       0       0
#> MAGEB2         0       0       0       0       0       0       0       0
#> SLC25A1        0       0       0       0       0       0       0       0
#> TFIP11         0       0       0       0       0       0       0       0
#> FBXO7          0       0       0       0       0       0       0       0
#> RPL3           0       0       0       0       0       0       0       0
#> EIF5           0       0       0       0       0       0       0       0
#> APEX1          0       0       0       0       0       0       0       0
#> RABGGTA        0       0       0       0       0       0       0       0
#> SALL4          0       0       0       0       0       0       0       0
#> ADNP           0       0       0       0       0       0       0       0
#> PSMA7          0       0       0       0       0       0       0       0
#> MRGBP          0       0       0       0       0       0       0       0
#> ASB9           0       0       0       0       0       0       0       0
#> RBBP7          0       0       0       0       0       0       0       0
#> USP11          0       0       0       0       0       0       0       0
#> CHRAC1         0       0       0       0       0       0       0       0
#> LIN7B          0       0       0       0       0       0       0       0
#> GPI            0       0       0       0       0       0       0       0
#> GARS           0       0       0       0       0       0       0       0
#> SETX           0       0       0       0       0       0       0       0
#> ZFAND5         0       0       0       0       0       0       0       0
#> TFAM           0       0       0       0       0       0       0       0
#> RPL19          0       0       0       0       0       0       0       0
#> FTSJ3          0       0       0       0       0       0       0       0
#> VAT1           0       0       0       0       0       0       0       0
#> UNC119         0       0       0       0       0       0       0       0
#> GRPEL1         0       0       0       0       0       0       0       0
#> HSPA8          0       0       0       0       0       0       0       0
#> DDX6           0       0       0       0       0       0       0       0
#> CBL            0       0       0       0       0       0       0       0
#> CAPRIN2        0       0       0       0       0       0       0       0
#> KRR1           0       0       0       0       0       0       0       0
#> TPI1           0       0       0       0       0       0       0       0
#> TDP2           0       0       0       0       0       0       0       0
#> ZNF451         0       0       0       0       0       0       0       0
#> RPS12          0       0       0       0       0       0       0       0
#> ARRDC3         0       0       0       0       0       0       0       0
#> RARS           0       0       0       0       0       0       0       0
#> ACTR8          0       0       0       0       0       0       0       0
#> EEF1B2         0       0       0       0       0       0       0       0
#> ID2            0       0       0       0       0       0       0       0
#> ODC1           0       0       0       0       0       0       0       0
#> CACYBP         0       0       0       0       0       0       0       0
#> LRRC42         0       0       0       0       0       0       0       0
#> PHF13          0       0       0       0       0       0       0       0
#> ADPRHL2        0       0       0       0       0       0       0       0
#> EXOC8          0       0       0       0       0       0       0       0
#> TXNIP          0       0       0       0       0       0       0       0
#> CCNI           0       0       0       0       0       0       0       0
#> ALG2           0       0       0       0       0       0       0       0
#> NEK9           0       0       0       0       0       0       0       0
#> IRF2BPL        0       0       0       0       0       0       0       0
#> TMEM214        0       0       0       0       0       0       0       0
#> YPEL5          0       0       0       0       0       0       0       0
#> FAM98A         0       0       0       0       0       0       0       0
#> AFTPH          0       0       0       0       0       0       0       0
#> LGALSL         0       0       0       0       0       0       0       0
#> MRPS14         0       0       0       0       0       0       0       0
#> NFYB           0       0       0       0       0       0       0       0
#> DPPA4          0       0       0       0       0       0       0       0
#> NAA50          0       0       0       0       0       0       0       0
#> FYTTD1         0       0       0       0       0       0       0       0
#> NUDT10         0       0       0       0       0       0       0       0
#> EXOSC9         0       0       0       0       0       0       0       0
#> RAB22A         0       0       0       0       0       0       0       0
#> SOX4           0       0       0       0       0       0       0       0
#> C17orf53       0       0       0       0       0       0       0       0
#>          SmExp39 SmExp40 SmExp41   SmExp42 SmExp43   SmExp44    SmExp45 SmExp46
#> TCEB3          0       0       0  0.339861       0  0.811890 -1.5047780       0
#> CLK1           0       0       0  2.345447       0  2.998250  1.9597690       0
#> MATR3          0       0       0  1.359442       0 -0.793046  0.1639185       0
#> AKAP11         0       0       0  1.030525       0  0.969406  0.7549080       0
#> HSF2           0       0       0  0.454356       0  0.032219  1.6180630       0
#> TBPL1          0       0       0  1.628324       0 -0.919032 -1.1052460       0
#> MAGEC2         0       0       0  2.612191       0 -2.605884  0.9727690       0
#> DDX20          0       0       0  2.166378       0  1.635119  3.4079575       0
#> ZBTB11         0       0       0  0.247616       0 -1.143104 -0.5803920       0
#> ARFGEF1        0       0       0  1.002807       0 -0.567846  1.2173180       0
#> FGFR3          0       0       0 -0.380591       0 -0.682199 -0.3717675       0
#> IKBKAP         0       0       0 -0.649878       0  0.095534  0.2233100       0
#> ASNS           0       0       0  0.612988       0 -0.336678  1.4813385       0
#> TCF3           0       0       0  0.128406       0  0.801963  0.1925195       0
#> PDCD2          0       0       0 -0.837151       0 -0.046529 -1.0470515       0
#> SMARCE1        0       0       0 -0.258172       0  0.802048 -0.6752805       0
#> LMAN1          0       0       0 -0.158695       0 -0.890028 -0.2087725       0
#> SLC25A3        0       0       0  1.985745       0  1.247226  1.3664045       0
#> REXO2          0       0       0 -1.175675       0 -1.027310  2.2162165       0
#> DDX43          0       0       0 -1.108005       0  2.366606  0.1350820       0
#> BZW1           0       0       0  1.216463       0  0.719985 -1.5697815       0
#> CCNT2          0       0       0  0.662712       0 -0.434991 -0.2933485       0
#> REST           0       0       0  1.334407       0  0.724851  1.0035830       0
#> MYNN           0       0       0 -0.150799       0 -0.875274 -1.0737755       0
#> ZNF213         0       0       0 -0.514661       0 -0.146602 -0.7339725       0
#> AKR1B1         0       0       0  3.181282       0 -0.998890 -0.7710735       0
#> SF3B2          0       0       0  0.784366       0  1.225847 -0.6015300       0
#> PABPC4         0       0       0 -0.177626       0  0.169746  0.7800320       0
#> CDV3           0       0       0  0.819141       0  1.254105  1.8996430       0
#> HECTD1         0       0       0  1.139056       0  0.854090  1.4841545       0
#> TDRD1          0       0       0  0.670724       0 -0.772104  0.0359300       0
#> BRPF3          0       0       0  1.359171       0  0.299340  0.2101305       0
#> TMEM14A        0       0       0 -0.406991       0  0.894532  0.1169825       0
#> MAGEB2         0       0       0  0.423041       0  2.167447  1.1658855       0
#> SLC25A1        0       0       0  0.858432       0 -0.727650 -0.8658545       0
#> TFIP11         0       0       0 -0.862156       0  1.557824  0.9116950       0
#> FBXO7          0       0       0  0.718727       0  1.459188 -0.1889275       0
#> RPL3           0       0       0  0.643272       0  0.617011 -0.1003200       0
#> EIF5           0       0       0  0.784738       0  0.510412 -0.0646560       0
#> APEX1          0       0       0  1.749741       0  4.083486  0.2756900       0
#> RABGGTA        0       0       0 -1.407364       0  0.228702 -0.5687040       0
#> SALL4          0       0       0  2.471999       0  3.546128  2.2771720       0
#> ADNP           0       0       0  0.371385       0  1.440851  1.3837595       0
#> PSMA7          0       0       0  0.063236       0 -1.236969  1.2043280       0
#> MRGBP          0       0       0 -1.088573       0  2.030355  0.0485285       0
#> ASB9           0       0       0  2.022701       0 -0.727417 -0.3993950       0
#> RBBP7          0       0       0 -0.151462       0 -0.958740 -0.6845300       0
#> USP11          0       0       0 -0.644672       0  1.930414 -0.1017215       0
#> CHRAC1         0       0       0  1.736766       0 -0.548720  0.5957730       0
#> LIN7B          0       0       0  1.294709       0 -1.050616  2.4947415       0
#> GPI            0       0       0  1.961681       0 -0.964023  0.3475470       0
#> GARS           0       0       0 -1.268631       0  2.033718  1.4896860       0
#> SETX           0       0       0 -0.861014       0  0.889896 -1.0164155       0
#> ZFAND5         0       0       0  0.159559       0  2.808798  1.8686630       0
#> TFAM           0       0       0 -0.272284       0 -0.010144  1.2491385       0
#> RPL19          0       0       0 -0.419557       0  1.361143  0.6682515       0
#> FTSJ3          0       0       0  0.852619       0 -0.167503  0.1082930       0
#> VAT1           0       0       0 -0.871333       0 -0.721666 -0.1343510       0
#> UNC119         0       0       0  0.663468       0  2.179575 -0.9035300       0
#> GRPEL1         0       0       0  0.822091       0  3.068360  0.0653905       0
#> HSPA8          0       0       0  0.691073       0  0.155760 -0.3968325       0
#> DDX6           0       0       0  0.415029       0  0.966323  0.4587580       0
#> CBL            0       0       0  0.348904       0 -0.609327 -0.4082835       0
#> CAPRIN2        0       0       0  0.491512       0  1.298165 -0.4603780       0
#> KRR1           0       0       0  1.711321       0 -1.074840  1.6460490       0
#> TPI1           0       0       0  2.030386       0  0.608469 -0.4174065       0
#> TDP2           0       0       0  1.922934       0 -1.012979  0.7376255       0
#> ZNF451         0       0       0  0.467404       0 -0.450371 -1.1181530       0
#> RPS12          0       0       0  0.436355       0  0.509580  0.8804555       0
#> ARRDC3         0       0       0 -1.189589       0  3.083748 -0.5493565       0
#> RARS           0       0       0 -1.114335       0  1.636677 -0.5915885       0
#> ACTR8          0       0       0  0.592046       0 -0.525401  0.1877385       0
#> EEF1B2         0       0       0  0.734650       0  0.422614  0.9752670       0
#> ID2            0       0       0 -0.813959       0  1.077503 -0.9012920       0
#> ODC1           0       0       0 -1.948449       0  2.033826  0.3300435       0
#> CACYBP         0       0       0  1.115591       0  0.902835  0.9587130       0
#> LRRC42         0       0       0  1.683990       0  0.801891  0.1180665       0
#> PHF13          0       0       0 -0.886886       0  2.951974  0.1982985       0
#> ADPRHL2        0       0       0 -1.018957       0 -0.907789 -1.1062910       0
#> EXOC8          0       0       0 -0.978517       0  3.308852 -1.0658505       0
#> TXNIP          0       0       0  2.272524       0  2.855201 -0.3797910       0
#> CCNI           0       0       0  0.161278       0 -0.604372  0.7137860       0
#> ALG2           0       0       0 -0.766728       0 -0.655559 -0.8540610       0
#> NEK9           0       0       0 -0.207997       0 -0.158411 -0.5508520       0
#> IRF2BPL        0       0       0  1.238072       0  1.139084 -0.8112455       0
#> TMEM214        0       0       0  0.512237       0  1.373279  0.0348805       0
#> YPEL5          0       0       0  0.781566       0 -2.718690  0.1578585       0
#> FAM98A         0       0       0  1.916290       0  2.512111  1.4767060       0
#> AFTPH          0       0       0  1.219302       0  0.412236  0.0668485       0
#> LGALSL         0       0       0  1.351143       0 -1.132657  1.1081695       0
#> MRPS14         0       0       0  1.295520       0 -0.757712 -0.4138390       0
#> NFYB           0       0       0 -0.867394       0  0.260824  1.6125925       0
#> DPPA4          0       0       0  0.544339       0 -0.587344  0.9645885       0
#> NAA50          0       0       0  0.128925       0  1.226955  2.4218010       0
#> FYTTD1         0       0       0 -0.925777       0 -0.602201  0.6747225       0
#> NUDT10         0       0       0 -0.637877       0 -0.526709 -0.5716210       0
#> EXOSC9         0       0       0  1.797679       0 -1.097839 -1.0639185       0
#> RAB22A         0       0       0  1.109180       0  1.754814  1.9880725       0
#> SOX4           0       0       0  1.246622       0  1.376727  0.6246885       0
#> C17orf53       0       0       0  2.474875       0 -0.335314 -0.0884200       0
#>          SmExp47 SmExp48 SmExp49   SmExp50
#> TCEB3          0       0       0 -1.693274
#> CLK1           0       0       0 -2.276759
#> MATR3          0       0       0  2.346283
#> AKAP11         0       0       0 -2.200790
#> HSF2           0       0       0  4.372338
#> TBPL1          0       0       0  4.821119
#> MAGEC2         0       0       0  0.294975
#> DDX20          0       0       0 -0.407157
#> ZBTB11         0       0       0 -1.501500
#> ARFGEF1        0       0       0  0.089197
#> FGFR3          0       0       0 -0.176738
#> IKBKAP         0       0       0  2.202534
#> ASNS           0       0       0  3.200386
#> TCF3           0       0       0  2.874046
#> PDCD2          0       0       0 -1.235547
#> SMARCE1        0       0       0  1.262509
#> LMAN1          0       0       0 -1.397875
#> SLC25A3        0       0       0 -2.621776
#> REXO2          0       0       0 -0.141224
#> DDX43          0       0       0 -1.351972
#> BZW1           0       0       0 -1.763238
#> CCNT2          0       0       0  1.763236
#> REST           0       0       0  2.738290
#> MYNN           0       0       0 -1.262271
#> ZNF213         0       0       0 -0.922468
#> AKR1B1         0       0       0  1.872678
#> SF3B2          0       0       0 -2.822344
#> PABPC4         0       0       0  2.100310
#> CDV3           0       0       0  2.511623
#> HECTD1         0       0       0  2.314945
#> TDRD1          0       0       0 -1.418241
#> BRPF3          0       0       0 -1.160660
#> TMEM14A        0       0       0 -1.148896
#> MAGEB2         0       0       0  2.470718
#> SLC25A1        0       0       0 -1.172708
#> TFIP11         0       0       0 -1.103456
#> FBXO7          0       0       0  0.097730
#> RPL3           0       0       0 -0.743368
#> EIF5           0       0       0  0.145704
#> APEX1          0       0       0 -2.178112
#> RABGGTA        0       0       0 -1.683193
#> SALL4          0       0       0 -0.987876
#> ADNP           0       0       0  3.104477
#> PSMA7          0       0       0 -1.339234
#> MRGBP          0       0       0  2.644684
#> ASB9           0       0       0 -1.137615
#> RBBP7          0       0       0  0.622404
#> USP11          0       0       0  0.001722
#> CHRAC1         0       0       0 -1.337750
#> LIN7B          0       0       0  2.860275
#> GPI            0       0       0  0.413296
#> GARS           0       0       0 -1.766900
#> SETX           0       0       0 -2.108651
#> ZFAND5         0       0       0 -0.989122
#> TFAM           0       0       0  1.951045
#> RPL19          0       0       0 -0.611346
#> FTSJ3          0       0       0 -1.347352
#> VAT1           0       0       0 -0.058669
#> UNC119         0       0       0 -1.092026
#> GRPEL1         0       0       0 -1.463594
#> HSPA8          0       0       0 -2.369458
#> DDX6           0       0       0 -0.212869
#> CBL            0       0       0  1.395326
#> CAPRIN2        0       0       0 -2.083040
#> KRR1           0       0       0 -1.378874
#> TPI1           0       0       0  0.281290
#> TDP2           0       0       0 -1.492390
#> ZNF451         0       0       0 -0.426896
#> RPS12          0       0       0  2.262664
#> ARRDC3         0       0       0  3.582882
#> RARS           0       0       0  4.652769
#> ACTR8          0       0       0 -0.939308
#> EEF1B2         0       0       0  1.648198
#> ID2            0       0       0 -0.071664
#> ODC1           0       0       0 -1.216449
#> CACYBP         0       0       0  2.833315
#> LRRC42         0       0       0 -0.877364
#> PHF13          0       0       0 -1.996727
#> ADPRHL2        0       0       0 -1.294786
#> EXOC8          0       0       0 -1.254346
#> TXNIP          0       0       0 -2.733702
#> CCNI           0       0       0  1.998889
#> ALG2           0       0       0 -0.546138
#> NEK9           0       0       0 -1.056071
#> IRF2BPL        0       0       0 -0.063241
#> TMEM214        0       0       0 -1.039049
#> YPEL5          0       0       0  2.530194
#> FAM98A         0       0       0 -1.325932
#> AFTPH          0       0       0 -0.880353
#> LGALSL         0       0       0 -1.573852
#> MRPS14         0       0       0 -1.144709
#> NFYB           0       0       0 -1.226326
#> DPPA4          0       0       0 -0.974342
#> NAA50          0       0       0 -0.977773
#> FYTTD1         0       0       0  2.856228
#> NUDT10         0       0       0 -0.913706
#> EXOSC9         0       0       0  2.458841
#> RAB22A         0       0       0 -1.301219
#> SOX4           0       0       0  0.407977
#> C17orf53       0       0       0 -1.115484