Get candidate genes/TFs-related peaks

refine peaks through genes in expression profile of scRNA-seq data

get_related_peaks(list1, expression_profile)

Arguments

list1
list, generated by get_related_genes(), the first element should be bed data.frame(contain 3 columns), the second element should contain 9 columns

expression_profile
dataframe, indicating expression profile of bulk RNA-seq or single-cell RNA-seq.

Value

return a list contain two dataframe, first one is candidate genes/TFs-related peaks dataframe, second one is bed dataframe

Examples

load(system.file("extdata", "list1.rda", package = "IReNA"))
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
Kmeans_clustering_ENS <- add_ENSID(test_clustering, Spec1='Hs')
list1 <- get_related_peaks(list1,Kmeans_clustering_ENS)

#> [1] "Get candidate genes/TFs-related peaks"