Function to visualize intramodular regulatory network

This function integrate R package igraph to visualize consequence of regulatory network analysis

plot_intramodular_network(
  TFs_list,
  enrichment = NULL,
  layout = "circle",
  group.cols = NULL,
  title.name = NULL,
  vertex.size = 25,
  vertex.label.color = "black",
  edge.label.color = "black",
  legend = TRUE,
  vertex.label.cex = 1.1,
  vertex.label.family = "ArialMT",
  frame.color = "white",
  arrow.size = 0.5,
  arrow.width = 0.8,
  edge.width = 2.5,
  edge.curved = 0,
  edge.color = c("#FDD1B0", "#B3B3B3"),
  vertex.label.degree = 0.35,
  vertex.label.dist = -6
)

Arguments

TFs_list
TFs_list generated by network_analysis

enrichment
Enrichment analysis results of genes in each module, generated by enrich_module. If it is not NULL, when plotting the Intramodular Network, the functions enriched by each module will be plotted in the diagram

layout
the layout to display the network, options: 'grid','sphere','circle','random'

group.cols
colors for group in network

title.name
the name of the title

vertex.size
vertex size

vertex.label.color
ertex label color

edge.label.color
edge label color

legend
logic, indicating whether to show the legend

vertex.label.cex
The label size of vertex

vertex.label.family
vertex label family

frame.color
frame.color

arrow.size
arrow size

arrow.width
arrow width

edge.width
edge width

edge.curved
edge curvature

edge.color
edge color. You need to input two colors, first one indicate 'Positive' regulation, second one indicate 'Negative' regulation.

vertex.label.degree
The position of the label in relation to the vertex (This parameter is valid only when the enrichment does not equal null)

vertex.label.dist
Distance between the label and the vertex (This parameter is valid only when the enrichment does not equal null)

Value

figure

Examples

#'
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
Kmeans_clustering <- add_ENSID(test_clustering, Spec1 = "Hs")
cor0.6 <- get_cor(Kmeans_clustering, Tranfac201803_Hs_MotifTFsF, 0.6, start_column=3)

#> [1] "get correlation for gene pairs"

TFs_list <- network_analysis(cor0.6,Kmeans_clustering)

#> [1] "Total TFs: 214"
#> [1] "Enriched TFs: 175"
#> [1] "Significant regulations: 6"

#enrichment <- enrich_module(Kmeans_cluster_Ens, org.Hs.eg.db, 'KEGG')
#plot_intramodular_network(TFs_list,enrichment,layout = 'random')